+++ CHANGELOG.md
@@ -26,6 +26,7 @@ QML Material CHANGELOG
  * [Documentation] Fix qdoc warnings
  * [PageStack] Pop pages when the Android back button is tapped
  * [Popup] Support preventing a popup from being dismissed
+ * [ActionBar] Control the overflow menu with Android hardware keys
 
 
 ### Version 0.0.6 (April 16, 2015)